+++ libc/src/__support/OSUtil/linux/vdso.h
@@ -7,43 +7,11 @@
 //===----------------------------------------------------------------------===//
 #ifndef LLVM_LIBC_SRC___SUPPORT_OSUTIL_LINUX_VDSO_H
 #define LLVM_LIBC_SRC___SUPPORT_OSUTIL_LINUX_VDSO_H
-#include "hdr/types/clock_t.h"
-#include "hdr/types/clockid_t.h"
-#include "hdr/types/struct_timespec.h"
-#include "hdr/types/struct_timeval.h"
-#include "hdr/types/time_t.h"
 #include "src/__support/CPP/array.h"
 #include "src/__support/common.h"
 #include "src/__support/macros/attributes.h"
 #include "src/__support/macros/properties/architectures.h"
 #include "src/__support/threads/callonce.h"
-
-// NOLINTBEGIN(llvmlibc-implementation-in-namespace)
-// TODO: some of the following can be defined via proxy headers.
-struct __kernel_timespec;
-struct timezone;
-struct riscv_hwprobe;
-struct getcpu_cache;
-struct cpu_set_t;
-// NOLINTEND(llvmlibc-implementation-in-namespace)
-
-namespace LIBC_NAMESPACE_DECL {
-namespace vdso {
-
-enum class VDSOSym {
-  ClockGetTime,
-  ClockGetTime64,
-  GetTimeOfDay,
-  GetCpu,
-  Time,
-  ClockGetRes,
-  RTSigReturn,
-  FlushICache,
-  RiscvHwProbe,
-  VDSOSymCount
-};
-} // namespace vdso
-} // namespace LIBC_NAMESPACE_DECL
 
 #if defined(LIBC_TARGET_ARCH_IS_X86)
 #include "x86_64/vdso.h"
@@ -59,34 +27,6 @@ enum class VDSOSym {
 
 namespace LIBC_NAMESPACE_DECL {
 namespace vdso {
-
-template <VDSOSym sym> LIBC_INLINE constexpr auto dispatcher() {
-  if constexpr (sym == VDSOSym::ClockGetTime)
-    return static_cast<int (*)(clockid_t, timespec *)>(nullptr);
-  else if constexpr (sym == VDSOSym::ClockGetTime64)
-    return static_cast<int (*)(clockid_t, __kernel_timespec *)>(nullptr);
-  else if constexpr (sym == VDSOSym::GetTimeOfDay)
-    return static_cast<int (*)(timeval *__restrict, timezone *__restrict)>(
-        nullptr);
-  else if constexpr (sym == VDSOSym::GetCpu)
-    return static_cast<int (*)(unsigned *, unsigned *, getcpu_cache *)>(
-        nullptr);
-  else if constexpr (sym == VDSOSym::Time)
-    return static_cast<time_t (*)(time_t *)>(nullptr);
-  else if constexpr (sym == VDSOSym::ClockGetRes)
-    return static_cast<int (*)(clockid_t, timespec *)>(nullptr);
-  else if constexpr (sym == VDSOSym::RTSigReturn)
-    return static_cast<void (*)(void)>(nullptr);
-  else if constexpr (sym == VDSOSym::FlushICache)
-    return static_cast<void (*)(void *, void *, unsigned int)>(nullptr);
-  else if constexpr (sym == VDSOSym::RiscvHwProbe)
-    return static_cast<int (*)(riscv_hwprobe *, size_t, size_t, cpu_set_t *,
-                               unsigned)>(nullptr);
-  else
-    return static_cast<void *>(nullptr);
-}
-
-template <VDSOSym sym> using VDSOSymType = decltype(dispatcher<sym>());
 
 class Symbol {
   VDSOSym sym;
@@ -115,21 +55,22 @@ private:
   static VDSOArray global_cache;
   static void initialize_vdso_global_cache();
 
-  template <typename T> LIBC_INLINE T get() {
+  LIBC_INLINE void *get() const {
     if (name().empty() || !is_valid())
       return nullptr;
 
     callonce(&once_flag, Symbol::initialize_vdso_global_cache);
-    return cpp::bit_cast<T>(global_cache[*this]);
+    return (global_cache[*this]);
   }
   template <VDSOSym sym> friend struct TypedSymbol;
 };
 
 template <VDSOSym sym> struct TypedSymbol {
   LIBC_INLINE constexpr operator VDSOSymType<sym>() const {
-    return Symbol{sym}.get<VDSOSymType<sym>>();
+    return cpp::bit_cast<VDSOSymType<sym>>(Symbol{sym}.get());
   }
-  template <typename... Args> LIBC_INLINE auto operator()(Args &&...args) {
+  template <typename... Args>
+  LIBC_INLINE auto operator()(Args &&...args) const {
     return this->operator VDSOSymType<sym>()(cpp::forward<Args>(args)...);
   }
 };
